requirements.txt
numpy
matplotlib
scipy
gTTS
librosa
soundfile
Pillow